refactor: move settings
